# backend/openapi_baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PhantomNet Sentinel & Cyber Defense API",
    "description": "PhantomNet V3 -- AI-Driven Active Defense, Automated Incident Response Playbooks, MITRE ATT&CK Mapping, IDS Rule Generation, and TAXII 2.1 Threat Sharing Platform",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/management/register": {
      "post": {
        "tags": [
          "Management"
        ],
        "summary": "Register Node",
        "operationId": "register_node_api_v1_management_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/management/heartbeat": {
      "post": {
        "tags": [
          "Management"
        ],
        "summary": "Node Heartbeat",
        "operationId": "node_heartbeat_api_v1_management_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/management/nodes": {
      "get": {
        "tags": [
          "Management"
        ],
        "summary": "List Nodes",
        "operationId": "list_nodes_api_v1_management_nodes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/management/policies": {
      "get": {
        "tags": [
          "Management"
        ],
        "summary": "List Policies",
        "operationId": "list_policies_api_v1_management_policies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "Management"
        ],
        "summary": "Create Policy",
        "operationId": "create_policy_api_v1_management_policies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/management/policies/assign": {
      "post": {
        "tags": [
          "Management"
        ],
        "summary": "Assign Policy",
        "operationId": "assign_policy_api_v1_management_policies_assign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/api/v1/hunting/search": {
      "post": {
        "tags": [
          "Threat Hunting"
        ],
        "summary": "Search Events",
        "operationId": "search_events_api_v1_hunting_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvancedQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hunting/extract-iocs": {
      "post": {
        "tags": [
          "Threat Hunting"
        ],
        "summary": "Extract Iocs",
        "operationId": "extract_iocs_api_v1_hunting_extract_iocs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IOCOutput"
                  },
                  "type": "array",
                  "title": "Response Extract Iocs Api V1 Hunting Extract Iocs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hunting/related-events": {
      "get": {
        "tags": [
          "Threat Hunting"
        ],
        "summary": "Get Related Events",
        "operationId": "get_related_events_api_v1_hunting_related_events_get",
        "parameters": [
          {
            "name": "ip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ip"
            }
          },
          {
            "name": "honeypot",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50
                },
                {
                  "type": "null"
                }
              ],
              "title": "Honeypot"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 43200,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "default": 1440,
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hunting/history": {
      "get": {
        "tags": [
          "Threat Hunting"
        ],
        "summary": "Get Search History",
        "operationId": "get_search_history_api_v1_hunting_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchHistoryResponse"
                  },
                  "title": "Response Get Search History Api V1 Hunting History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hunting/analyze-patterns": {
      "post": {
        "tags": [
          "Threat Hunting"
        ],
        "summary": "Analyze Patterns",
        "operationId": "analyze_patterns_api_v1_hunting_analyze_patterns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvancedQuery": {
        "properties": {
          "logic": {
            "type": "string",
            "title": "Logic",
            "default": "AND"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/QueryCondition"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Conditions",
            "default": []
          },
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "AdvancedQuery"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IOCOutput": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "type",
          "value"
        ],
        "title": "IOCOutput"
      },
...
      "NodeHeartbeatRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Node Id"
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "NodeHeartbeatRequest"
      },
      "NodeRegisterRequest": {
        "properties": {
          "hostname": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Hostname"
          },
          "ip_address": {
            "type": "string",
            "maxLength": 45,
            "minLength": 1,
            "title": "Ip Address"
          },
          "honeypot_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Honeypot Type"
          }
        },
        "type": "object",
        "required": [
          "hostname",
          "ip_address",
          "honeypot_type"
        ],
        "title": "NodeRegisterRequest"
      },
...
      "PolicyAssignRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Node Id"
          },
          "policy_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Policy Id"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "policy_id"
        ],
        "title": "PolicyAssignRequest"
      },
      "PolicyCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "title": "Description"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "config"
        ],
        "title": "PolicyCreateRequest"
      },
      "QueryCondition": {
        "properties": {
          "field": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Field"
          },
          "operator": {
            "type": "string",
            "title": "Operator"
          },
          "value": {
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "field",
          "operator",
          "value"
        ],
        "title": "QueryCondition"
      },
...
      "SearchHistoryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "query_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Json"
          },
          "result_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Count",
            "default": 0
          },
          "executed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Name"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "SearchHistoryResponse"
      },
...
      "TextPayload": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500000,
            "title": "Text",
            "description": "Raw log or text payload"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TextPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}